[feature/S09P22A110-439] 챌린지의 accumulatedAmount Integer->Long 수정

See merge request s09-fintech-finance-sub2/S09P22A110!97

Former-commit-id: a0878282e98182ad4a4191d8b68839daac70aef8

+++ akgimi/src/main/java/com/kangkimleekojangcho/akgimi/challenge/application/response/CreateChallengeServiceResponse.java
@@ -9,14 +9,14 @@ import java.time.LocalDate;
 public class CreateChallengeServiceResponse {
     private final Long challengeId;
     private final Long productId;
-    private final Integer accumulatedAmount;
+    private final Long accumulatedAmount;
     private final boolean achievementState;
     private final Integer challengePeriod;
     private final LocalDate challengeStartDate;
     private final LocalDate challengeEndDate;
     private final boolean isInProgress;
 
-    private CreateChallengeServiceResponse(Long challengeId, Long productId, Integer accumulatedAmount, boolean achievementState, Integer challengePeriod, LocalDate challengeStartDate, LocalDate challengeEndDate, boolean isInProgress) {
+    private CreateChallengeServiceResponse(Long challengeId, Long productId, Long accumulatedAmount, boolean achievementState, Integer challengePeriod, LocalDate challengeStartDate, LocalDate challengeEndDate, boolean isInProgress) {
         this.challengeId = challengeId;
         this.productId = productId;
         this.accumulatedAmount = accumulatedAmount;
